+++ RandomSkunk.Results.Dapper/GlobalUsings.cs
@@ -3,6 +3,5 @@ global using System;
 global using System.Collections.Generic;
 global using System.Data;
 global using System.Data.Common;
-global using System.Reflection;
 global using System.Runtime.InteropServices;
 global using System.Threading.Tasks;